Handle formatting changes between files
Adds logic and tests for line-by-line format handling.  Lines with format
changes will be reformated as default (unquoted), unless the reformatting is
the only change and is applied on both sides of the merge.

Lines where there is no content changes and LCA and A sides match, will
preserve that formatting.