Add relevance-based ordering logic for 3-way line conflicts.
When LCA, A and B all have different next keys, use relevance distance to try
to keep next matching keys as close as possible when deciding whether to emit
the A or B key.  See notes.txt for more on the algorithm.

Also adds tests for asymmetric, overlapping addition/deletion of lines on A/B
sides; ensure that output lines are in the correct order even when A and B are
swapped.