Improve output of merge conflicts.
Adds the line key to the line identifier when introducing a conflicting line;

Turns newlines into "\n" strings when output an individual field conflict, to
ensure that each conflict remains in a single line with an appropriate conflict
prefix.  Especially if the row's input text is split over multiple lines, we
need to be very clear what lines are conflict field markers and what lines
represent the actual input file content.