Add --reformat-all option to force reformatting of output lines
Normally we preserve the input lines exactly, including irrelevant formatting
details such as optional field quotations, when input lines match and there are
no conflicts.  We reformat output lines only when we need to change the content.

The --reformat-all option forces all output lines to be reformatted, allowing
the user to force a specific formatting style (eg. --quote-all) on the output.