Handle lines missing a primary key entirely.
It is not uncommon to have blank lines in a CSV file (eg. a separator line
after the initial column names in the first row.)  Such lines may simply not have a primary key.

Add a method to lookup a field in a Line by column, giving a default if that
column is not found; and use that to default the primary key to the empty
string for Lines missing that column.

Also adds a test for adding blank lines to an existing file.