Fix handling of blank lines.
A blank line will have an empty string as its key; this is a special case but
should still work.

Make sure that the debug code handles this correctly by falling back on "" as
the key if we cannot find a field in the line at the key index.

And in the merge logic, make sure we test for the existence of a key by using
"key ==/!= None" rather than just "key", as the latter will fail if the key is
the empty string.  (The empty string evaluates as False when cast to a
boolean.)