Fix one further case of blank line handling
If we find a line that is different in all three branches (ie. the LCA key is
different from both A and B keys), but it's not a delete, then we know the LCA
key must exist in some future line in both A and B: hence neither side can be
empty.

We test for this by asserting that both the A and B keys are present when
handling this 3-way-diff state.  But the assert was just checking for the key
evaluating to True, which breaks if the key is present but is an empty string.

So test explicitly that the key != None instead.