Add a random test generator/fuzzer.
Generates virtual CSV files using custom RNG-driven generators, and perturbs
these at random to introduce various effects in the A and B branches to
exercise the merge.  Effects include moving, deleting, cloning or altering
lines, and adding lines with blank keys.

Intended to be run continuously for fuzz testing, not as part of normal unit
tests.