Initial abstraction of merge3 output handling.
Creates an abstract OutputDriver class, with Merge3OutputDriver as the subclass
which implements 3-way merge output.

This is the first step towards adding 2- and 3-way diff output, which will be
implemented as alternative OutputDriver sub-classes.

The initial version here is enough to pass all existing tests, but we do not
yet pass data into the OutputDriver in enough detail to reliably construct
alternative output views, so further refactoring is expected.