Further OutputDriver refactoring.
Add the Line objects for LCA, A and B to each output method for the
OutputDriver.  The diff2/diff3 drivers will be able to use these to determine
which lines have been deleted, added or modified on each side of the diff.

As a result, we won't need distinct methods for merged/shared lines; we can
tell which lines are shared between which files by the line_* arguments.