Pass through even deleted lines to the output driver
If a line is deleted, pass it through to the OutputDriver's emit_text method
with a None text line.  We will ignore it for 3-way merge, but diff output
methods will be able to use this to mark deleted lines.  The absence of line_A
and/or line_B will give information as to which side of the merge has deleted
the line.