Always emit deleted lines
The diff2 OutputDriver needs to see all deleted lines, but we were not calling
the necessary emit function in the case where a key in the LCA has been
completely deleted from both A and B.

Add that emit_text() call, supplying no text.  It will still be ignored for
merge3 but makes sure diff catches the delete.