Allow a key to exist multiple times in the backlog
If we are reordering a line which appears multiple times, then it is possible
for that line to need to reside multiple times in the backlog.

So turn backlog into an orderedmultidict, and add/remove single values from the
multidict's list at a time.

Also cleans up the debug dumping of the backlog to match.