Add diff preamble and git diff support
Adds support in 2-way csvdiff for diff preamble in the output, ie.
preamble lines of the form

    csvdiff -k"key" file1.csv file2.csv
    --- file1.csv
    +++ file2.csv

Additionally, adds a git-aware mode that accepts arguments in the format emitted by
git to a GIT_EXTERNAL_DIFF driver, and formats the preamble appropriately for git.

It should be possible to set up a git diff driver with eg.

    [diff "csvdiff"]
   	command = csvdiff_git -k\"key\" --colour
	renames = false

and obtain useful git output without need for an additional wrapper script to
handle the 7-fold arguments that git will supply to the command.