Add Key string subclass to represent line keys
We rely on Key values being None to represent a key being entirely
missing within a given file.

But we cannot just test this with "if key:", as another special case --- the
empty key, "" --- also evaluates to False; such keys commonly arise if there is
eg. a blank line in the csv file.

So add a new Key class, as a subclass of strings, which always evaluates
to True when cast to boolean.  That way the empty string will not trigger
tests for a missing key.

Also updates the Line class's get_field method to always return a Key
object, and to always use a Key('') as the default for a missing field.

Simplifies some tests/asserts to depend on the new logic.