Add option for 2-way diff to show reordered lines
Adds the ability to pass output-driver-specific args to the OutputDriver classes

Adds the --show-reordered-lines option to csvdiff

Adds a test for lines which are reordered but not otherwise changed during
csvdiff output (ie. lines which are passed to emit_text(), not emit_row()): if
all the lines being emitted are at the current cursor positions in the input
files, then the lines are in the expected order so we skip them.  Otherwise,
the output has been reordered, and we still show it if the show_reordered_lines
option is set.