Initial refactoring of line-order merge code
There are many different cases to handle when we find unmatched, out-of-order
lines: inserts, deletes, moves forward and backward, pushing to backlog and
matching against backlog etc.

The old logic was spread over several functions, making it hard to reason about
exactly what cases are handled in which order, and what assumptions we can make
as we eliminate different possibilities.

So refactor the various cases handled in resync_LCA() and sub-functions, into a
single, long but well-commented set of tests in merge3_next().

In the process, fixes a long-standing issue where on certain reordering events,
we would push an LCA line to the backlog without also pushing a matching line
in A or B.

This is a precursor to other changes to make the handling of duplicate keys
more robust.