Improve robustness of backlog handling with duplicate keys
When we have multiple lines with the same key, handling reorder is particularly
complex.

In particular, consider the case where LCA has two lines with the same key; one
of those lines gets reordered (in either A, B or both) to later on in the file,
and the second line is deleted altogether.

In this case:
* The first LCA instance matches later on in (A and/or B) and is pushed to backlog
* Before we come across the matching A/B line ---
  * the *second* LCA instance appears, and we look forward for its match.

If we are not careful, we will find the same match and conclude that the second
LCA line is also moved, so we push it to the backlog; we will error out at the
EOF when we find that the backlog is not empty, as there was never a second
line coming to consume that second instance in the backlog.

So:

* Mark matching lines in both A and B as "has_backlog_match" when we push an
  LCA line to the backlog
* Skip over such lines when we search for a potential match for any subsequent
  LCA lines with the same key
* Also record the exact A/B matches we determine when we push an LCA line to
  the backlog, and
* Use those recorded matches when eventually emitting the merged lines; we must
  not repeat a search for the matching lines at that point as we will already have
  marked the matches as "has_backlog_match" (hence they will be skipped in such
  searches.)