Only use cached state when emitting backlog match lines
When we have multiple lines with the same key, it is really important that we
match lines between files consistently.

So remove the final uses of match searching when we finally emit a line that
was previously matched with another line from the LCA which was moved to the
backlog.

Rather,

* when we move lines to the backlog, record all the LCA/A/B matches we
  determine at that time;

* Use only that cached info when we finally reach the matching line later on
  that triggers an emit;

* Always skip over backlog lines in find_next_match: once they have been found
  and matched once, we never want to match them against a different line.