Relax expectations in Cursor.consume()
When we consume a line, it may not always be the next instance of that line's key in the file.

That will *usually* be the case, but it falls down when:

* we have multiple lines with the same key,
* one is pushed to the backlog, and
* another one is deleted

In this case, the deleted line may get emitted (and hence consumed) before we
get to the backlog line; so the backlog will contain a still-valid, un-consumed
line with the same key, that is earlier in the file than the line we're deleting.