Add line numbers in debug output
Add line numbers to the default string format method for the Line class, so
that when we write lines to debug output (in eg. merge_one_line()), we always
include the exact line number.

Useful when debugging files containing duplicate keys/lines, where just seeing
the line contents is not always enough.