Use a predictable common key rather than '' in fuzz testing
When the random_test fuzz tester randomly perturbs a file, one option it has is
to change the key to a common fixed value to maximise the number of line
duplicates we get with that key throughout the file.

Previously we just used the empty string as the key value in this case.  Change
that to the fixed string 'common', as this is much easier to search for and
parse in debug log files.