Check backlog matches before handling unmoved/reordered lines.
For performance, we try to check whether the next line in each file has the
same key, and can therefore be quickly matched, before we do more complex tests
for insert/delete/reorder etc.

But if we have already determined that a line is a match against the backlog,
we need to perform that match in preference --- when we have multiple lines
with the same key, we always match them in strict order, so if a key has
moved/disappeared in one file, we may have a backlog match even if we later on
find the keys in the same position.

We must always preserve the existing backlog match in that case.

Eg.

 LCA           A             B
 name,value    name,value    name,value
 foo,1         foo,1         foo,1
 common,2
 bar,3         bar,3         bar,3
 common,4      common,4      common,4

When we hit LCA "common,2", we push that to backlog after first matching
against common,4 in A and B.  This looks counter-intuitive but is a consequence
of strict order matching for duplicates.  So in this case, when we get to
"common,4", we *must* first match A and B against the LCA common,2 backlog
line; so we test that backlog match *before* testing whether the keys match on
all 3 lines.

A future improvement may be to identify that "common,2" is actually deleted in
this case above.  But we cannot actually determine that without first looking
later on in the file; eg. if "common,4" were not present in LCA, then that case
would actually represent a move of the single "common" line.  So that decision
can only be made if we are able to defer decisions about which lines are
reordered vs. being deleted, which makes backlog behaviour less predictable
(eg. it is no longer a failure to have an empty backlog at the end of a file).
So we make a deliberate decision to start with this strict in-order matching to
verify correctness with that more predictable behaviour, before extending the
code to handle other duplicate matching strategies.