Handle renamed files in csvdiff_git
Git expects external diff drivers to accept two additional optional arguments
if the source file has been renamed: the new name of the file, and a set of
optional additional text to display in the diff header describing the rename.

Accept this additional text in the csvdiff_git CLI entrypoint, pass it in to
the diff2 OutputDriver, and display it unconditionally in the diff header if
present.