CSVFile: enable loading of empty files
Allows us to load completely empty files without erroring out on the
absence of a header line.

On load of such files, we insert a dummy EmptyLine as the header
contents to facilitate lookup of the set of columns present (which will
be empty too.)

EOF() always returns true on an empty file.