Enable 2-way diff of changed columns
Handling added/removed columns is already performed robustly by the core
merge3 code, but we were not handling all cases of this in the 2-way
diff output driver.

Improve that in the driver's emit_csv_row() method, which is called any
time we have a line where the output includes some field conflicts
between the old and new versions (in this case, a new/removed field).
We need to handle the cases where the field is entirely missing in the
old or new version of the line.

The header lines are already output with this method.  Rather than
looking up the name of that line by primary key as we do for most lines,
label the header line's diff output manually with a special "<Column
names>" chunk in the output, requiring a row-name override to be
accepted by the emit_csv_row() method generally; eg.

@@ -1 +1 @@ <Column names>
 {+Id+},{+Name+}

for a new file with those keys.