Headers: add support for full output key maps.
The Headers class is responsible for figuring out which columns need to
appear in the output, given the headers of each input file.

In 3-way merge, that output list necessarily drops any columns which
have been deleted on either side of the merge (the A or B file).

But for diffs, we want to explicitly annotate deleted data rather than
just dropping it silently.  So we need an additional output map that
contains *all* the keys from any of the input files, including both
added and deleted columns.  This requires reinserting deleted keys back
into the header map.

Add a map_all_headers() method to do precisely this.  It will determine
all visible column keys, identify the ones missing from the merged output
map, and construct a new map with those missing keys reinsterted in a
reasonable location.