Fix quoting of lines in 2-way diff output
We need to quote content in 2-way diff output (eg. newline characters
are printed as "\n" to avoid confusing line splits in the output.)

When emitting a line of output wholesale (because eg. that line has been
added or deleted in its entirety), we would previously pass that line to
the field-by-field emit function for this, faking the "difference" by
providing the entire line as both before and after image.

But this fails if the columns have changed between the before and after
files; we'll potentially look up the content in the wrong location if we
pass a line in as if it were from a different file.

So just quote each field directly when doing a whole-line emit; it's
simpler and safer.