Headers: cache the full header output map
We previously only every needed a full consolidated header output map
for one case: csvdiff2, where we need to know all possible header keys
from all input files, including ones deleted by the diff, in order to
format that delation in the diff text.

But we need that same output map when calculating automatic/guessed
primary keys, so we may need this calculation for merge3 too in certain
cases.

So allow header.map_all_headers() to be called multiple times: cache its
result the first time it is used and return that if it is called again.