Primary key guessing: handle empty files.
If all input files are empty, then we expect to produce an empty diff
output, regardless of the primary key selected (every primary key is
compatible with empty input files.)

But [auto] primary key guessing was failing to find a key in this case,
even though the rest of the diff logic can handle it.

So handle this by generating an automatic, easily-identifiable fake
primary key when guessing keys if all input files are empty.